test: restructure modularised db tests
