--- tests/test_dbapi.py
@@ -1,1553 +1,0 @@
-import importlib
-from datetime import datetime, timedelta
-
-import pytest
-import sqlalchemy as sa
-from alembic.migration import MigrationContext
-from alembic.operations import Operations
-
-from tahrir_api.model import (
-    Assertion,
-    Authorization,
-    Badge,
-    Milestone,
-    Person,
-    Rarity,
-    Series,
-    Tag,
-    Team,
-)
-
-
-@pytest.fixture
-def dummy_issuer_id(api):
-    return api.add_issuer("TestOrigin", "TestName", "TestOrg", "TestContact")
-
-
-@pytest.fixture
-def dummy_badge_id(api, dummy_issuer_id):
-    return api.add_badge(
-        "TestBadge",
-        "TestImage",
-        "A test badge for doing unit tests",
-        "TestCriteria",
-        dummy_issuer_id,
-    )
-
-
-@pytest.fixture
-def dummy_person_id(api):
-    return api.add_person("test@tester.com")
-
-
-@pytest.fixture
-def initialize_list_assertions(api, dummy_person_id, dummy_issuer_id):
-    for unit in range(0, 10):
-        tmpbadge = api.add_badge(
-            f"AsrtName_{unit}",
-            f"AsrtShot_{unit}",
-            f"AsrtDesc_{unit}",
-            f"AsrtCrit_{unit}",
-            dummy_issuer_id,
-        )
-        api.add_assertion(tmpbadge, "test@tester.com", None, f"link_{unit}")
-    return 10
-
-
-@pytest.fixture
-def search_test_badges(api, dummy_issuer_id):
-    api.add_badge(
-        name="Python Expert",
-        image="python_expert.png",
-        desc="Badge for Python programming excellence",
-        criteria="Complete 100 Python exercises",
-        issuer_id=dummy_issuer_id,
-        tags=["python", "programming", "expert"],
-    )
-    api.add_badge(
-        name="JavaScript Ninja",
-        image="js_ninja.png",
-        desc="Master the JavaScript language and frameworks",
-        criteria="Build 5 JavaScript projects",
-        issuer_id=dummy_issuer_id,
-        tags=["javascript", "web", "frontend"],
-    )
-    api.add_badge(
-        name="Frontend Master",
-        image="frontend.png",
-        desc="Expert web developer with full-stack capabilities",
-        criteria="Complete web development bootcamp",
-        issuer_id=dummy_issuer_id,
-        tags=["web", "development", "fullstack"],
-    )
-    api.add_badge(
-        name="Doc Writer",
-        image="doc_writer.png",
-        desc="Created comprehensive project documentation",
-        criteria="Write 50 pages of documentation",
-        issuer_id=dummy_issuer_id,
-        tags=["documentation", "writing", "communication"],
-    )
-    api.add_badge(
-        name="OSS Contributor",
-        image="oss_contributor.png",
-        desc="Active contributor to open source projects and python ecosystems",
-        criteria="Contribute to 3 open source projects",
-        issuer_id=dummy_issuer_id,
-        tags=["opensource", "python", "community"],
-    )
-    return 5
-
-
-def test_add_badges(api, dummy_badge_id):
-    assert api.get_badge("testbadge").__str__() == "TestBadge"
-    assert api.badge_exists("testbadge") is True
-
-
-def test_add_badge_creates_tag_relationships(api, dummy_issuer_id):
-    badge_id = api.add_badge(
-        "TaggedBadge",
-        "tagged.png",
-        "A badge with normalized tags",
-        "Tag criteria",
-        dummy_issuer_id,
-        tags=["python", "testing"],
-    )
-
-    badge = api.get_badge(badge_id)
-
-    assert {tag.name for tag in badge.tags} == {"python", "testing"}
-    assert len(badge.tags) == 2
-
-
-def test_add_badge_reuses_existing_tags(api, dummy_issuer_id):
-    api.add_badge(
-        "FirstTaggedBadge",
-        "first.png",
-        "First badge with a shared tag",
-        "First criteria",
-        dummy_issuer_id,
-        tags=["shared", "first"],
-    )
-    api.add_badge(
-        "SecondTaggedBadge",
-        "second.png",
-        "Second badge with a shared tag",
-        "Second criteria",
-        dummy_issuer_id,
-        tags=["shared", "second"],
-    )
-
-    assert api.session.query(Tag).filter_by(name="shared").count() == 1
-
-
-def test_create_series_creates_tag_relationships(api):
-    team_id = api.create_team("SeriesTeam")
-
-    series_id = api.create_series(
-        "TaggedSeries",
-        "A series with normalized tags",
-        team_id,
-        tags="python, testing",
-    )
-
-    series = api.get_series(series_id)
-
-    assert {tag.name for tag in series.tags} == {"python", "testing"}
-    assert len(series.tags) == 2
-
-
-def test_create_series_deduplicates_comma_separated_tags(api):
-    team_id = api.create_team("DedupSeriesTeam")
-
-    series_id = api.create_series(
-        "DedupTaggedSeries",
-        "A series with duplicate tags",
-        team_id,
-        tags="python, testing, python,  testing  ,",
-    )
-
-    series = api.get_series(series_id)
-
-    assert {tag.name for tag in series.tags} == {"python", "testing"}
-    assert api.session.query(Tag).filter_by(name="python").count() == 1
-
-
-def test_add_team(api):
-    api.create_team("TestTeam")
-
-    assert api.team_exists("testteam") is True
-
-
-def test_add_series(api):
-    team_id = api.create_team("TestTeam")
-
-    api.create_series("TestSeries", "A test series", team_id, "test, series")
-
-    assert api.series_exists("testseries") is True
-
-
-def test_add_milestone(api, dummy_issuer_id):
-    team_id = api.create_team("TestTeam")
-    series_id = api.create_series("TestSeries", "A test series", team_id, "test, series")
-
-    badge_id_1 = api.add_badge(
-        "TestBadge-1",
-        "TestImage-2",
-        "A test badge for doing 10 unit tests",
-        "TestCriteria",
-        dummy_issuer_id,
-    )
-
-    badge_id_2 = api.add_badge(
-        "TestBadge-2",
-        "TestImage-2",
-        "A test badge for doing 100 unit tests",
-        "TestCriteria",
-        dummy_issuer_id,
-    )
-
-    milestone_id_1 = api.create_milestone(1, badge_id_1, series_id)
-
-    milestone_id_2 = api.create_milestone(2, badge_id_2, series_id)
-
-    assert api.milestone_exists(milestone_id_1) is True
-    assert api.milestone_exists(milestone_id_2) is True
-
-
-def test_add_person(api, dummy_person_id):
-    assert api.get_person("test@tester.com").__str__() == "test@tester.com"
-    assert api.person_exists("test@tester.com") is True
-
-
-def test_add_issuer(api, dummy_issuer_id):
-    assert api.get_issuer(dummy_issuer_id).__str__() == "TestName"
-    assert api.issuer_exists("TestOrigin", "TestName") is True
-
-
-def test_add_invitation(api, dummy_badge_id, dummy_person_id):
-    _id = api.add_invitation(dummy_badge_id, created_by_email="test@tester.com")
-
-    assert api.invitation_exists(_id)
-    invitation = api.get_invitation(_id)
-    assert api.get_person(id=invitation.created_by).email == "test@tester.com"
-
-
-def test_add_invitation_no_created_by(api, dummy_badge_id, dummy_person_id):
-    with pytest.raises(ValueError):
-        api.add_invitation(dummy_badge_id)
-
-
-def test_add_invitation_rejects_legacy_badge(api, dummy_badge_id, dummy_person_id):
-    """Test that add_invitation rejects creating an invitation for a legacy badge"""
-    badge = api.get_badge(dummy_badge_id)
-    badge.legacy = True
-    api.session.flush()
-
-    with pytest.raises(ValueError, match="legacy badge"):
-        api.add_invitation(dummy_badge_id, created_by_email="test@tester.com")
-
-
-def test_expire_invitation(api, dummy_badge_id, dummy_person_id):
-    # Create
-    future_time = datetime.now() + timedelta(hours=2)
-    _id = api.add_invitation(
-        dummy_badge_id, created_by_email="test@tester.com", expires_on=future_time
-    )
-
-    # Verify
-    assert api.invitation_exists(_id)
-    invitation = api.get_invitation(_id)
-    assert not invitation.expired
-
-    # Expire
-    result = api.expire_invitation(_id)
-    assert result is True
-
-    # Verify
-    expired_invitation = api.get_invitation(_id)
-    assert expired_invitation.expired
-
-    # Absent
-    result = api.expire_invitation("non-existent-id")
-    assert result is False
-
-
-def test_last_login(api, callback_calls, dummy_person_id):
-    person = api.get_person(dummy_person_id)
-    assert not person.last_login
-    api.note_login(nickname=person.nickname)
-    assert person.last_login
-
-    assert len(callback_calls) == 1
-    message = callback_calls[0][0][0]
-    assert message.body == {"user": {"username": "test", "badges_user_id": 1}}
-    assert message.agent_name == "test"
-    assert message.summary == "test logged into badges for the first time"
-
-
-def test_add_assertion(api, callback_calls, dummy_badge_id, dummy_person_id):
-    api.add_assertion(dummy_badge_id, "test@tester.com", None, "link")
-    assert api.assertion_exists(dummy_badge_id, "test@tester.com")
-
-    badge = api.get_badge(dummy_badge_id)
-    assert badge.assertions[0].issued_for == "link"
-    assert api.get_assertions_by_badge(dummy_badge_id)[0].__str__() == "TestBadge<->test@tester.com"
-
-    # Ensure that we would have published a fedmsg messages for that.
-    assert len(callback_calls) == 1
-
-    award_message = callback_calls[0][0][0]
-
-    assert award_message.body == {
-        "badge": {
-            "name": "TestBadge",
-            "description": "A test badge for doing unit tests",
-            "image_url": "TestImage",
-            "badge_id": "testbadge",
-        },
-        "user": {"username": "test", "badges_user_id": 1},
-    }
-    assert award_message.agent_name is None
-    assert award_message.usernames == ["test"]
-    assert award_message.summary == "test was awarded the badge `TestBadge`"
-
-
-def test_add_assertion_rejects_legacy_badge(api, dummy_badge_id, dummy_person_id):
-    """Test that add_assertion rejects awarding a legacy badge"""
-    badge = api.get_badge(dummy_badge_id)
-    badge.legacy = True
-    api.session.flush()
-    with pytest.raises(ValueError, match="legacy badge"):
-        api.add_assertion(dummy_badge_id, "test@tester.com", None)
-    assert api.assertion_exists(dummy_badge_id, "test@tester.com") is False
-
-
-@pytest.mark.parametrize("test_email", ["test@tester.com", "Test@Tester.Com"])
-def test_get_assertions_by_email(api, callback_calls, dummy_badge_id, dummy_person_id, test_email):
-    api.add_assertion(dummy_badge_id, "test@tester.com", None, "link")
-    # This should be case-insensitive
-    assertions = api.get_assertions_by_email(person_email=test_email)
-    assert len(assertions) == 1
-    assert str(assertions[0]) == "TestBadge<->test@tester.com"
-
-
-def test_adjust_ranks(api, callback_calls, dummy_badge_id, dummy_person_id):
-    person = api.get_person(dummy_person_id)
-    assertion = Assertion(
-        badge_id=dummy_badge_id,
-        person_id=person.id,
-        issued_for="link",
-    )
-    api.session.add(assertion)
-    api.session.flush()
-
-    api.adjust_ranks(person)
-
-    # Ensure that we would have published a fedmsg messages for that.
-    assert len(callback_calls) == 1
-    rank_advance_message = callback_calls[0][0][0]
-    assert rank_advance_message.body == {
-        "person": {
-            "email": "test@tester.com",
-            "id": 1,
-            "nickname": "test",
-            "website": None,
-            "bio": None,
-            "rank": 1,
-        },
-        "old_rank": None,
-    }
-    assert rank_advance_message.agent_name is None
-    assert rank_advance_message.usernames == ["test"]
-    assert rank_advance_message.summary == "test's Badges rank changed from None to 1"
-
-
-def test_get_badges_from_tags(api, dummy_issuer_id):
-    # Badge tagged with "test"
-    api.add_badge(
-        "TestBadgeA",
-        "TestImage",
-        "A test badge for doing unit tests",
-        "TestCriteria",
-        dummy_issuer_id,
-        tags=["test"],
-    )
-
-    # Badge tagged with "tester"
-    api.add_badge(
-        "TestBadgeB",
-        "TestImage",
-        "A second test badge for doing unit tests",
-        "TestCriteria",
-        dummy_issuer_id,
-        tags=["tester"],
-    )
-
-    # Badge tagged with both "test" and "tester"
-    api.add_badge(
-        "TestBadgeC",
-        "TestImage",
-        "A third test badge for doing unit tests",
-        "TestCriteria",
-        dummy_issuer_id,
-        tags=["test", "tester"],
-    )
-
-    tags = ["test", "tester"]
-    badges_any = api.get_badges_from_tags(tags, match_all=False)
-    assert len(badges_any) == 3
-    badges_all = api.get_badges_from_tags(tags, match_all=True)
-    assert len(badges_all) == 1
-
-
-def test_remove_assertion_success(api, dummy_badge_id, dummy_person_id):
-    """Test successful removal of an assertion"""
-    result = api.add_assertion(dummy_badge_id, "test@tester.com", None)
-    assert result == ("test@tester.com", dummy_badge_id)
-    assert api.assertion_exists(dummy_badge_id, "test@tester.com") is True
-
-    result = api.remove_assertion(dummy_badge_id, "test@tester.com")
-    assert result
-    assert api.assertion_exists(dummy_badge_id, "test@tester.com") is False
-
-    assertions = api.get_assertions_by_email("test@tester.com")
-    assert len(assertions) == 0
-
-
-def test_remove_assertion_nonexistent_person(api, dummy_badge_id):
-    """Test removing assertion for non-existent person returns False"""
-    result = api.remove_assertion(dummy_badge_id, "nonexistent@example.com")
-    assert result is False
-
-
-def test_remove_assertion_nonexistent_badge(api, dummy_person_id):
-    """Test removing assertion for non-existent badge returns False"""
-    result = api.remove_assertion("nonexistent-badge", "test@tester.com")
-    assert result is False
-
-
-def test_remove_assertion_nonexistent_assertion(api, dummy_badge_id, dummy_person_id):
-    """Test removing non-existent assertion returns False"""
-    result = api.remove_assertion(dummy_badge_id, "test@tester.com")
-    assert result is False
-
-
-@pytest.mark.parametrize(
-    "kwargs",
-    [
-        # Test individual field updates
-        {"name": "UpdatedName"},
-        {"image": "UpdatedImage"},
-        {"description": "Updated description"},
-        {"criteria": "Updated criteria"},
-        {"tags": ["updated", "tags"]},
-        {"tags": ["updated", "tags"]},
-        # Test multiple field updates
-        {
-            "name": "MultiUpdate",
-            "image": "MultiImage",
-            "description": "Multi description",
-            "criteria": "Multi criteria",
-            "tags": ["multi", "tags"],
-        },
-        # Test empty update
-        {},
-    ],
-)
-def test_update_badge(api, dummy_badge_id, kwargs):
-    """Test updating badge"""
-    # Obtain
-    existing_badge = api.get_badge(dummy_badge_id)
-
-    # Create
-    expected_name = kwargs.get("name", existing_badge.name)
-    expected_image = kwargs.get("image", existing_badge.image)
-    expected_description = kwargs.get("description", existing_badge.description)
-    expected_criteria = kwargs.get("criteria", existing_badge.criteria)
-
-    # Handle
-    if "tags" in kwargs:
-        expected_tags = set(kwargs["tags"])
-    else:
-        expected_tags = {tag.name for tag in existing_badge.tags}
-
-    # Update
-    api.update_badge(dummy_badge_id, **kwargs)
-
-    # Obtain
-    updated_badge = api.get_badge(dummy_badge_id)
-
-    # Verify
-    assert updated_badge.name == expected_name
-    assert updated_badge.image == expected_image
-    assert updated_badge.description == expected_description
-    assert updated_badge.criteria == expected_criteria
-    assert {tag.name for tag in updated_badge.tags} == expected_tags
-
-
-def test_update_badge_replaces_tags(api, dummy_issuer_id):
-    badge_id = api.add_badge(
-        "ReplaceTagsBadge",
-        "replace.png",
-        "A badge whose tags will be replaced",
-        "Replace criteria",
-        dummy_issuer_id,
-        tags=["old", "shared"],
-    )
-
-    api.update_badge(badge_id, tags=["new", "shared"])
-
-    updated_badge = api.get_badge(badge_id)
-    assert {tag.name for tag in updated_badge.tags} == {"new", "shared"}
-    assert api.session.query(Tag).filter_by(name="shared").count() == 1
-
-
-def test_update_badge_nonexistent(api):
-    """Test updating badge which are non-existent and returns False"""
-    result = api.update_badge("nonexistent_badge_id", name="UpdatedName")
-    assert result is False
-
-
-def test_update_badge_invalid_fields(api, dummy_badge_id):
-    """Test updating badge with invalid fields and raises KeyError"""
-    with pytest.raises(KeyError, match="Invalid fields"):
-        api.update_badge(dummy_badge_id, invalid_field="value")
-
-
-@pytest.mark.parametrize("legacy_value", [True, False])
-def test_update_badge_legacy_field(api, dummy_badge_id, legacy_value):
-    """Test that update_badge can modify the legacy status of a badge"""
-    api.update_badge(dummy_badge_id, legacy=legacy_value)
-    badge = api.get_badge(dummy_badge_id)
-    assert badge.legacy is legacy_value
-
-
-@pytest.mark.parametrize(
-    "identifier_type,update_data",
-    [
-        ("email", {"avatar": "test@test-libravatar.com"}),
-        ("id", {"bio": "Updated bio via ID"}),
-        ("nickname", {"website": "https://nickname-example.com"}),
-    ],
-)
-def test_update_person_by_identifier(api, dummy_person_id, identifier_type, update_data):
-    """Test updating person profile fields using different identification methods."""
-    person = api.get_person("test@tester.com")
-    assert person.website is None
-    assert person.bio is None
-    assert person._avatar is None
-
-    if identifier_type == "email":
-        update_args = {"person_email": "test@tester.com"}
-    elif identifier_type == "id":
-        update_args = {"id": person.id}
-    else:
-        update_args = {"nickname": person.nickname}
-
-    lookup_args = update_args.copy()
-    update_args.update(update_data)
-    result = api.update_person(**update_args)
-    assert result.email == "test@tester.com"
-
-    # Verify the updated fields are present in the returned object
-    for dictname, expected in update_data.items():
-        formname = "_avatar" if dictname == "avatar" else dictname
-        assert getattr(result, formname) == expected
-
-    # Also verify by fetching fresh from database
-    updated_person = api.get_person(**lookup_args)
-    for field, expected_value in update_data.items():
-        actual_field = "_avatar" if field == "avatar" else field
-        assert getattr(updated_person, actual_field) == expected_value
-
-
-def test_update_person_none_values(api, dummy_person_id):
-    """Test that None values don't overwrite existing data."""
-    # Set initial values
-    api.update_person(
-        person_email="test@tester.com",
-        website="https://example.com",
-        bio="Existing bio",
-        avatar="test@test-libravatar.com",
-    )
-    result = api.update_person(person_email="test@tester.com", website=None, bio=None, avatar=None)
-
-    # Verify the returned object still has the original values (not overwritten by None)
-    assert result.email == "test@tester.com"
-    assert result.website == "https://example.com"
-    assert result.bio == "Existing bio"
-    assert result._avatar == "test@test-libravatar.com"
-
-    # Also verify by fetching fresh from database
-    person = api.get_person("test@tester.com")
-    assert person.website == "https://example.com"
-    assert person.bio == "Existing bio"
-    assert person._avatar == "test@test-libravatar.com"
-
-
-@pytest.mark.parametrize(
-    "identifier_type,identifier_value",
-    [("email", "nonexistent@example.com"), ("id", 999999), ("nickname", "nonexistent_user")],
-)
-def test_update_person_nonexistent(api, identifier_type, identifier_value):
-    """Test updating a person that doesn't exist should return False."""
-    if identifier_type == "email":
-        update_args = {"person_email": identifier_value}
-    elif identifier_type == "id":
-        update_args = {"id": identifier_value}
-    else:
-        update_args = {"nickname": identifier_value}
-
-    update_args["website"] = "https://example.com"
-    result = api.update_person(**update_args)
-    assert result is False
-
-
-@pytest.mark.parametrize(
-    "begin, limit, expected_count",
-    [
-        (None, None, 10),
-        (0, 5, 5),
-        (5, 5, 5),
-        (0, 20, 10),
-        (5, None, 5),
-        (5, 10, 5),
-        (10, 10, 0),
-        (None, 0, 0),
-        (0, 1, 1),
-    ],
-)
-def test_get_all_assertions(api, initialize_list_assertions, begin, limit, expected_count):
-    """Test get_all_assertions with various begin and limit parameters."""
-    assertions = list(api.get_all_assertions(begin=begin, limit=limit))
-
-    assert len(assertions) == expected_count
-    if len(assertions) > 1:
-        assert assertions[0].issued_on >= assertions[1].issued_on
-
-    if begin is not None and begin > 0 and expected_count > 0:
-        first_batch = list(api.get_all_assertions(begin=0, limit=expected_count))
-        if len(first_batch) == expected_count:
-            assert assertions[0].id != first_batch[0].id
-
-
-def test_delete_authorization_success(api, dummy_badge_id, dummy_person_id):
-    """Test successful deletion of an authorization"""
-    person = api.session.query(Person).filter_by(email="test@tester.com").one()
-    authorization = Authorization(badge_id=dummy_badge_id, person_id=person.id)
-    api.session.add(authorization)
-    api.session.flush()
-
-    auth_count = (
-        api.session.query(Authorization)
-        .filter_by(badge_id=dummy_badge_id, person_id=person.id)
-        .count()
-    )
-    assert auth_count == 1
-
-    result = api.delete_authorization(dummy_badge_id, "test@tester.com")
-
-    auth_count = (
-        api.session.query(Authorization)
-        .filter_by(badge_id=dummy_badge_id, person_id=person.id)
-        .count()
-    )
-
-    assert result == ("test@tester.com", dummy_badge_id)
-    assert auth_count == 0
-
-
-def test_delete_authorization_nonexistent_person(api, dummy_badge_id):
-    """Test deleting authorization for non-existent person returns False"""
-    result = api.delete_authorization(dummy_badge_id, "nonexistent@example.com")
-    assert result is False
-
-
-def test_delete_authorization_nonexistent_authorization(api, dummy_badge_id, dummy_person_id):
-    """Test deleting non-existent authorization returns False"""
-    person = api.session.query(Person).filter_by(email="test@tester.com").one()
-    auth_count = (
-        api.session.query(Authorization)
-        .filter_by(badge_id=dummy_badge_id, person_id=person.id)
-        .count()
-    )
-    assert auth_count == 0
-
-    result = api.delete_authorization(dummy_badge_id, "test@tester.com")
-    assert result is False
-
-
-@pytest.mark.parametrize(
-    "badge_id,email,should_succeed",
-    [
-        ("testbadge", "test@tester.com", True),
-        ("testbadge", "nonexistent@example.com", False),
-        ("nonexistent-badge", "test@tester.com", False),
-    ],
-)
-def test_add_authorization(api, dummy_badge_id, dummy_person_id, badge_id, email, should_succeed):
-    """Test adding authorization with various badge and person combinations"""
-    result = api.add_authorization(badge_id, email)
-
-    if should_succeed:
-        assert result == (email, badge_id)
-        assert api.authorization_exists(badge_id, email) is True
-    else:
-        assert result is False
-        assert api.authorization_exists(badge_id, email) is False
-
-
-def test_add_authorization_rejects_legacy_badge(api, dummy_badge_id, dummy_person_id):
-    """Test that add_authorization rejects authorizing a legacy badge"""
-
-    badge = api.get_badge(dummy_badge_id)
-    badge.legacy = True
-    api.session.flush()
-
-    with pytest.raises(ValueError, match="legacy badge"):
-        api.add_authorization(dummy_badge_id, "test@tester.com")
-    assert api.authorization_exists(dummy_badge_id, "test@tester.com") is False
-
-
-def test_delete_badge_rejects_legacy_badge(api, dummy_badge_id):
-    """Test that delete_badge raises ValueError when the badge is marked as legacy"""
-    api.update_badge(dummy_badge_id, legacy=True)
-
-    with pytest.raises(ValueError, match="legacy badge"):
-        api.delete_badge(dummy_badge_id)
-
-    assert api.badge_exists(dummy_badge_id) is True
-
-
-def test_get_team(api):
-    """Test retrieving a team by ID"""
-    team = Team(id="test-team", name="Test Team")
-    api.session.add(team)
-    api.session.flush()
-
-    received_team = api.get_team("test-team")
-    assert received_team is not None
-    assert received_team.name == "Test Team"
-    assert received_team.id == "test-team"
-
-    absented_team = api.get_team("absented-team")
-    assert absented_team is None
-
-
-def test_get_series_from_team(api):
-    """Test retrieving series from a team"""
-    team = Team(id="test-team", name="Test Team")
-    api.session.add(team)
-    api.session.flush()
-
-    series = api.get_series_from_team("test-team")
-    assert series == []
-
-    series_a = Series(
-        id="test-series-alpha",
-        name="Test Series Alpha",
-        description="Alpha test series",
-        team_id="test-team",
-    )
-    series_b = Series(
-        id="test-series-bravo",
-        name="Test Series Bravo",
-        description="Bravo test series",
-        team_id="test-team",
-    )
-    api.session.add(series_a)
-    api.session.add(series_b)
-    api.session.flush()
-
-    series = api.get_series_from_team("test-team")
-    assert len(series) == 2
-
-    series_names = [s.name for s in series]
-    assert "Test Series Alpha" in series_names
-    assert "Test Series Bravo" in series_names
-
-    absented_series = api.get_series_from_team("absented-team")
-    assert absented_series is None
-
-
-def test_get_badges_from_team(api, dummy_issuer_id):
-    """Test retrieving badges from a team via series and milestones"""
-    team = Team(id="test-team", name="Test Team")
-    api.session.add(team)
-    api.session.flush()
-
-    badges = api.get_badges_from_team("test-team")
-    assert badges == []
-
-    series = Series(
-        id="test-series",
-        name="Test Series",
-        description="Test series",
-        team_id="test-team",
-    )
-    api.session.add(series)
-    api.session.flush()
-
-    badges = api.get_badges_from_team("test-team")
-    assert badges == []
-
-    badge_a = Badge(
-        id="test-badge-alpha",
-        name="Test Badge Alpha",
-        image="TestImageAlpha",
-        description="Test Badge Alpha",
-        criteria="TestCriteriaAlpha",
-        issuer_id=dummy_issuer_id,
-    )
-    badge_b = Badge(
-        id="test-badge-bravo",
-        name="Test Badge Bravo",
-        image="TestImageBravo",
-        description="Test Badge Bravo",
-        criteria="TestCriteriaBravo",
-        issuer_id=dummy_issuer_id,
-    )
-    api.session.add(badge_a)
-    api.session.add(badge_b)
-    api.session.flush()
-
-    milestone_a = Milestone(position=1, badge_id="test-badge-alpha", series_id="test-series")
-    milestone_b = Milestone(position=2, badge_id="test-badge-bravo", series_id="test-series")
-    api.session.add(milestone_a)
-    api.session.add(milestone_b)
-    api.session.flush()
-
-    badges = api.get_badges_from_team("test-team")
-    assert badges is not None
-    assert len(badges) == 2
-
-    badge_names = [b.name for b in badges]
-    assert "Test Badge Alpha" in badge_names
-    assert "Test Badge Bravo" in badge_names
-
-    absented_badges = api.get_badges_from_team("absented-team")
-    assert absented_badges is None
-
-
-def test_get_series_existing(api):
-    """Test getting an existing series by ID."""
-    team = Team(id="test-team", name="Test Team")
-    api.session.add(team)
-    api.session.flush()
-    series = Series(
-        id="test-series",
-        name="Test Series",
-        description="A test series for unit testing",
-        team_id="test-team",
-    )
-    api.session.add(series)
-    api.session.flush()
-
-    retrieved_series = api.get_series("test-series")
-    assert retrieved_series is not None
-    assert retrieved_series.id == "test-series"
-    assert retrieved_series.name == "Test Series"
-    assert retrieved_series.description == "A test series for unit testing"
-    assert retrieved_series.tags == []
-
-
-def test_get_series_nonexistent(api):
-    """Test getting a non-existent series returns None."""
-    series = api.get_series("absented-series")
-    assert series is None
-
-
-def test_get_all_series_empty(api):
-    """Test getting all series when none exist."""
-    direct_count = api.session.query(Series).count()
-    assert direct_count == 0
-    series_query = api.get_all_series()
-    series_list = list(series_query)
-    assert len(series_list) == 0
-
-
-@pytest.mark.parametrize(
-    "search_string, expected_total, expected_nicknames",
-    [
-        ("alice", 1, ["alice_wonder"]),
-        ("test", 2, ["dave_test", "diana_test"]),
-        ("zzznomatch", 0, []),
-        ("uppercase", 1, ["UpperCase"]),
-    ],
-)
-def test_get_persons_by_nickname_search(api, search_string, expected_total, expected_nicknames):
-    api.add_person("alice@test.com", nickname="alice_wonder")
-    api.add_person("dave@test.com", nickname="dave_test")
-    api.add_person("diana@test.com", nickname="diana_test")
-    api.add_person("upper@test.com", nickname="UpperCase")
-    result = api.get_persons_by_nickname(search_string)
-    assert result["total"] == expected_total
-    nicknames = [p.nickname for p in result["users"]]
-    assert nicknames == expected_nicknames
-
-
-@pytest.mark.parametrize(
-    "begin, limit, expected_count",
-    [
-        (0, 2, 2),
-        (2, 2, 2),
-        (4, 2, 1),
-    ],
-)
-def test_get_persons_by_nickname_pagination(api, begin, limit, expected_count):
-    for i in range(5):
-        api.add_person(f"user{i}@test.com", nickname=f"user_{i}")
-    result = api.get_persons_by_nickname("user", begin=begin, limit=limit)
-    assert len(result["users"]) == expected_count
-    assert result["total"] == 5
-    assert result["begin"] == begin
-    assert result["limit"] == limit
-
-
-@pytest.mark.parametrize(
-    "query, expected_total, must_contain",
-    [
-        ("Python Expert", 1, ["Python Expert"]),
-        ("python", 2, ["Python Expert", "OSS Contributor"]),
-        ("PYTHON", 2, ["Python Expert", "OSS Contributor"]),
-        ("web developer", 1, ["Frontend Master"]),
-        ("documentation", 1, ["Doc Writer"]),
-        ("nonexistent_badge_xyz", 0, []),
-        (
-            "",
-            5,
-            [
-                "Python Expert",
-                "OSS Contributor",
-                "Frontend Master",
-                "Doc Writer",
-                "JavaScript Ninja",
-            ],
-        ),
-    ],
-)
-def test_search_filter_conditions(query, expected_total, must_contain, api, search_test_badges):
-    result = api.get_badges_by_string(query)
-
-    assert result["total"] == expected_total
-    badge_names = [b.name for b in result["badges"]]
-    for name in must_contain:
-        assert name in badge_names
-
-
-@pytest.mark.parametrize(
-    "query, begin, limit, expected_begin, expected_limit, expected_count",
-    [
-        ("python", 0, 100, 0, 100, 2),
-        ("python", 0, 1, 0, 1, 1),
-        ("python", 1, 1, 1, 1, 1),
-        ("python", 500, 1, 500, 1, 0),
-        ("", 0, 500, 0, 100, 5),
-    ],
-)
-def test_search_pagination_conditions(
-    query, begin, limit, expected_begin, expected_limit, expected_count, api, search_test_badges
-):
-    result = api.get_badges_by_string(query, begin=begin, limit=limit)
-
-    assert result["limit"] == expected_limit
-    assert len(result["badges"]) == expected_count
-    assert result["total"] >= expected_count
-
-
-def test_badge_legacy_default(api, dummy_badge_id):
-    """Test that a newly created badge has legacy=False by default"""
-    badge = api.get_badge(dummy_badge_id)
-    assert badge.legacy is False
-
-
-@pytest.mark.parametrize(
-    "legacy_value,expected",
-    [
-        (False, False),
-        (True, True),
-    ],
-)
-def test_badge_legacy(api, dummy_badge_id, legacy_value, expected):
-    """Test that legacy can be set and retrieved correctly"""
-    badge = api.get_badge(dummy_badge_id)
-    badge.legacy = legacy_value
-    api.session.flush()
-
-    updated_badge = api.get_badge(dummy_badge_id)
-    assert updated_badge.legacy is expected
-
-
-@pytest.mark.parametrize(
-    "legacy_value,expected",
-    [
-        (False, False),
-        (True, True),
-    ],
-)
-def test_badge_legacy_in_as_dict(api, dummy_badge_id, legacy_value, expected):
-    """Test that as_dict includes the legacy field correctly"""
-    badge = api.get_badge(dummy_badge_id)
-    badge.legacy = legacy_value
-    api.session.flush()
-
-    badge_dict = api.get_badge(dummy_badge_id).as_dict()
-    assert "legacy" in badge_dict
-    assert badge_dict["legacy"] is expected
-
-
-@pytest.fixture
-def dummy_persons(api):
-    """Create 12 persons so bucket maths work cleanly with 12 badges."""
-    for i in range(12):
-        api.add_person(f"person{i}@test.com")
-
-
-@pytest.fixture
-def dummy_badges_with_assertions(api, dummy_issuer_id, dummy_persons):
-    """
-    Create 12 badges with descending assertion counts:
-    badge_0 → 12 assertions (most owned → D tier)
-    badge_11 → 1 assertion  (least owned → X tier)
-    """
-    for i in range(12):
-        badge_id = api.add_badge(
-            f"RarityBadge_{i}",
-            f"RarityImage_{i}",
-            f"RarityDesc_{i}",
-            f"RarityCrit_{i}",
-            dummy_issuer_id,
-        )
-        for j in range(12 - i):
-            api.add_assertion(badge_id, f"person{j}@test.com", None, f"link_{i}_{j}")
-
-
-def test_compute_badge_rarities_standard_case(api, dummy_badges_with_assertions):
-    """12 badges / 6 tiers: names, assignment, distribution, limits, X and D placement."""
-    api.compute_badge_rarities()
-
-    rarities = api.session.query(Rarity).all()
-    badges = api.get_all_badges().all()
-    rarity_by_name = {r.name: r for r in rarities}
-
-    assert len(rarities) == 6
-    assert set(rarity_by_name) == {"X", "S", "A", "B", "C", "D"}
-    assert all(badge.rarity_id is not None for badge in badges)
-
-    for rarity in rarities:
-        count = sum(1 for b in badges if b.rarity_id == rarity.id)
-        assert count == 2
-        assert rarity.lower_limit >= 0.0
-        assert rarity.upper_limit >= rarity.lower_limit
-
-    x_tier = rarity_by_name["X"]
-    d_tier = rarity_by_name["D"]
-    x_badges = [b for b in badges if b.rarity_id == x_tier.id]
-    d_badges = [b for b in badges if b.rarity_id == d_tier.id]
-    assert any("RarityBadge_11" in b.name for b in x_badges)
-    assert any("RarityBadge_0" in b.name for b in d_badges)
-
-
-def test_compute_badge_rarities_no_users_returns_early(api, dummy_issuer_id):
-    """With no users in the DB, compute should return without error."""
-    api.add_badge("EmptyBadge", "img", "desc", "crit", dummy_issuer_id)
-    api.compute_badge_rarities()  # should not raise
-    badges = api.get_all_badges().all()
-    assert all(badge.rarity_id is None for badge in badges)
-
-
-def test_compute_badge_rarities_idempotent(api, dummy_badges_with_assertions):
-    """Calling compute twice should update existing rows, not crash on unique constraints."""
-    api.compute_badge_rarities()
-    first_run = {r.name: (r.lower_limit, r.upper_limit) for r in api.session.query(Rarity).all()}
-
-    api.compute_badge_rarities()
-    second_run = {r.name: (r.lower_limit, r.upper_limit) for r in api.session.query(Rarity).all()}
-
-    assert api.session.query(Rarity).count() == 6
-    assert first_run == second_run
-
-
-def test_compute_badge_rarities_fewer_than_six_badges(api, dummy_issuer_id):
-    """With 3 badges and 6 tiers, some tiers should be empty."""
-    for i in range(3):
-        api.add_person(f"few{i}@test.com")
-    for i in range(3):
-        badge_id = api.add_badge(
-            f"FewBadge_{i}", f"img_{i}", f"desc_{i}", f"crit_{i}", dummy_issuer_id
-        )
-        for j in range(3 - i):
-            api.add_assertion(badge_id, f"few{j}@test.com", None, f"link_{i}_{j}")
-
-    api.compute_badge_rarities()
-    badges = api.get_all_badges().all()
-    assert all(b.rarity_id is not None for b in badges)
-
-
-def test_compute_badge_rarities_uneven_badge_count(api, dummy_issuer_id):
-    """7 badges / 6 tiers exercises the remainder distribution logic."""
-    for i in range(7):
-        api.add_person(f"uneven{i}@test.com")
-    for i in range(7):
-        badge_id = api.add_badge(
-            f"UnevenBadge_{i}", f"img_{i}", f"desc_{i}", f"crit_{i}", dummy_issuer_id
-        )
-        for j in range(7 - i):
-            api.add_assertion(badge_id, f"uneven{j}@test.com", None, f"link_{i}_{j}")
-
-    api.compute_badge_rarities()
-    badges = api.get_all_badges().all()
-    assert all(b.rarity_id is not None for b in badges)
-    tier_counts = {}
-    for b in badges:
-        tier_counts[b.rarity_id] = tier_counts.get(b.rarity_id, 0) + 1
-    assert sum(tier_counts.values()) == 7
-    # One tier gets the extra badge from remainder
-    assert max(tier_counts.values()) == 2
-
-
-def test_compute_badge_rarities_badge_with_zero_assertions(api, dummy_issuer_id):
-    """A badge with zero assertions should still be assigned a rarity."""
-    for i in range(6):
-        api.add_person(f"zero{i}@test.com")
-    for i in range(6):
-        badge_id = api.add_badge(
-            f"ZeroBadge_{i}", f"img_{i}", f"desc_{i}", f"crit_{i}", dummy_issuer_id
-        )
-        if i < 5:
-            api.add_assertion(badge_id, f"zero{i}@test.com", None, f"link_{i}")
-
-    api.compute_badge_rarities()
-    badges = api.get_all_badges().all()
-    assert all(b.rarity_id is not None for b in badges)
-    # The badge with 0 assertions should be in X tier (least owned)
-    zero_badge = next(b for b in badges if b.name == "ZeroBadge_5")
-    x_tier = api.session.query(Rarity).filter(Rarity.name == "X").first()
-    assert zero_badge.rarity_id == x_tier.id
-
-
-def test_badge_as_dict_includes_tag_names(api, dummy_issuer_id):
-    badge_id = api.add_badge(
-        "DictTagsBadge",
-        "dict-tags.png",
-        "A badge whose dict includes tag names",
-        "Dict criteria",
-        dummy_issuer_id,
-        tags=["dict", "tags"],
-    )
-
-    badge_dict = api.get_badge(badge_id).as_dict()
-
-    assert badge_dict["tags"] == ["dict", "tags"]
-
-
-def test_series_as_dict_includes_tag_names(api):
-    team_id = api.create_team("DictSeriesTeam")
-    series_id = api.create_series(
-        "DictTagsSeries",
-        "A series whose dict includes tag names",
-        team_id,
-        tags="dict, tags",
-    )
-
-    series_dict = api.get_series(series_id).as_dict()
-
-    assert series_dict["tags"] == ["dict", "tags"]
-
-
-series_tags_migration = importlib.import_module(
-    "tahrir_api.migrations.versions.196a305a2e0c_normalize_series_tags"
-)
-
-
-def run_series_tags_migration(connection, migration_func):
-    context = MigrationContext.configure(connection)
-    operations = Operations(context)
-
-    original_op = series_tags_migration.op
-    series_tags_migration.op = operations
-    try:
-        migration_func()
-    finally:
-        series_tags_migration.op = original_op
-
-
-def create_series_tags_old_schema(connection):
-    metadata = sa.MetaData()
-    sa.Table(
-        "series",
-        metadata,
-        sa.Column("id", sa.Unicode(length=128), primary_key=True),
-        sa.Column("tags", sa.Unicode(length=128), nullable=True),
-    )
-    sa.Table(
-        "tags",
-        metadata,
-        sa.Column("id", sa.Integer(), primary_key=True, autoincrement=True),
-        sa.Column("name", sa.Unicode(length=128), nullable=False, unique=True),
-    )
-    metadata.create_all(connection)
-
-
-def test_series_tags_upgrade_preserves_existing_tags():
-    engine = sa.create_engine("sqlite:///:memory:")
-
-    with engine.begin() as connection:
-        create_series_tags_old_schema(connection)
-        connection.execute(
-            sa.text(
-                """
-                INSERT INTO series (id, tags)
-                VALUES
-                    ('series-a', 'python, testing, python'),
-                    ('series-b', ' docs , writing ')
-                """
-            )
-        )
-
-        run_series_tags_migration(connection, series_tags_migration.upgrade)
-
-        series_columns = [column["name"] for column in sa.inspect(connection).get_columns("series")]
-        tag_rows = connection.execute(sa.text("SELECT name FROM tags ORDER BY name")).fetchall()
-        series_tag_rows = connection.execute(
-            sa.text(
-                """
-                SELECT series_tags.series_id, tags.name
-                FROM series_tags
-                JOIN tags ON tags.id = series_tags.tag_id
-                ORDER BY series_tags.series_id, tags.name
-                """
-            )
-        ).fetchall()
-
-    assert "tags" not in series_columns
-    assert [row[0] for row in tag_rows] == ["docs", "python", "testing", "writing"]
-    assert series_tag_rows == [
-        ("series-a", "python"),
-        ("series-a", "testing"),
-        ("series-b", "docs"),
-        ("series-b", "writing"),
-    ]
-
-
-def test_series_tags_downgrade_restores_comma_separated_tags():
-    engine = sa.create_engine("sqlite:///:memory:")
-
-    with engine.begin() as connection:
-        create_series_tags_old_schema(connection)
-        connection.execute(
-            sa.text("INSERT INTO series (id, tags) VALUES ('series-a', 'python, testing')")
-        )
-        run_series_tags_migration(connection, series_tags_migration.upgrade)
-
-        run_series_tags_migration(connection, series_tags_migration.downgrade)
-
-        series_columns = [column["name"] for column in sa.inspect(connection).get_columns("series")]
-        restored_tags = connection.execute(
-            sa.text("SELECT tags FROM series WHERE id = 'series-a'")
-        ).scalar()
-        tables = sa.inspect(connection).get_table_names()
-
-    assert "tags" in series_columns
-    assert restored_tags == "python, testing"
-    assert "series_tags" not in tables
-
-
-badge_tags_migration = importlib.import_module(
-    "tahrir_api.migrations.versions.1013e1694fa2_normalize_badge_tags"
-)
-
-
-def run_badge_tags_migration(connection, migration_func):
-    context = MigrationContext.configure(connection)
-    operations = Operations(context)
-
-    original_op = badge_tags_migration.op
-    badge_tags_migration.op = operations
-    try:
-        migration_func()
-    finally:
-        badge_tags_migration.op = original_op
-
-
-def create_badge_tags_old_schema(connection):
-    metadata = sa.MetaData()
-    sa.Table(
-        "badges",
-        metadata,
-        sa.Column("id", sa.Unicode(length=128), primary_key=True),
-        sa.Column("tags", sa.Unicode(length=128), nullable=True),
-    )
-    metadata.create_all(connection)
-
-
-def test_badge_tags_upgrade_preserves_existing_tags():
-    engine = sa.create_engine("sqlite:///:memory:")
-
-    with engine.begin() as connection:
-        create_badge_tags_old_schema(connection)
-        connection.execute(
-            sa.text(
-                """
-                INSERT INTO badges (id, tags)
-                VALUES
-                    ('badge-a', 'python, testing, python'),
-                    ('badge-b', ' docs , writing ')
-                """
-            )
-        )
-
-        run_badge_tags_migration(connection, badge_tags_migration.upgrade)
-
-        badge_columns = [column["name"] for column in sa.inspect(connection).get_columns("badges")]
-        tag_rows = connection.execute(sa.text("SELECT name FROM tags ORDER BY name")).fetchall()
-        badge_tag_rows = connection.execute(
-            sa.text(
-                """
-                SELECT badge_tags.badge_id, tags.name
-                FROM badge_tags
-                JOIN tags ON tags.id = badge_tags.tag_id
-                ORDER BY badge_tags.badge_id, tags.name
-                """
-            )
-        ).fetchall()
-
-    assert "tags" not in badge_columns
-    assert [row[0] for row in tag_rows] == ["docs", "python", "testing", "writing"]
-    assert badge_tag_rows == [
-        ("badge-a", "python"),
-        ("badge-a", "testing"),
-        ("badge-b", "docs"),
-        ("badge-b", "writing"),
-    ]
-
-
-def test_badge_tags_downgrade_restores_comma_separated_tags():
-    engine = sa.create_engine("sqlite:///:memory:")
-
-    with engine.begin() as connection:
-        create_badge_tags_old_schema(connection)
-        connection.execute(
-            sa.text("INSERT INTO badges (id, tags) VALUES ('badge-a', 'python, testing')")
-        )
-        run_badge_tags_migration(connection, badge_tags_migration.upgrade)
-
-        run_badge_tags_migration(connection, badge_tags_migration.downgrade)
-
-        badge_columns = [column["name"] for column in sa.inspect(connection).get_columns("badges")]
-        restored_tags = connection.execute(
-            sa.text("SELECT tags FROM badges WHERE id = 'badge-a'")
-        ).scalar()
-        tables = sa.inspect(connection).get_table_names()
-
-    assert "tags" in badge_columns
-    assert restored_tags == "python, testing"
-    assert "badge_tags" not in tables
-    assert "tags" not in tables
-
-
-def test_cleanup_orphan_tags(api, dummy_issuer_id):
-    api.add_badge(
-        "OrphanBadge",
-        "orphan.png",
-        "A badge whose tags may become orphans",
-        "Orphan criteria",
-        dummy_issuer_id,
-        tags=["unique-tag", "shared-tag"],
-    )
-    api.add_badge(
-        "KeeperBadge",
-        "keeper.png",
-        "A badge that keeps shared-tag alive",
-        "Keeper criteria",
-        dummy_issuer_id,
-        tags=["shared-tag", "keeper-tag"],
-    )
-
-    assert api.session.query(Tag).count() == 3
-
-    api.delete_badge("orphanbadge")
-    deleted = api.cleanup_orphan_tags()
-
-    assert deleted == 1
-    remaining = {t.name for t in api.session.query(Tag).all()}
-    assert "unique-tag" not in remaining
-    assert "shared-tag" in remaining
-    assert "keeper-tag" in remaining
-
-
-def test_delete_badge_with_tags(api, dummy_issuer_id):
-    api.add_badge(
-        "DeleteMe",
-        "delete.png",
-        "A badge to be deleted",
-        "Delete criteria",
-        dummy_issuer_id,
-        tags=["delete-tag", "common-tag"],
-    )
-    api.add_badge(
-        "KeepMe",
-        "keep.png",
-        "A badge to keep",
-        "Keep criteria",
-        dummy_issuer_id,
-        tags=["common-tag"],
-    )
-
-    result = api.delete_badge("deleteme")
-    assert result == "deleteme"
-    assert api.badge_exists("deleteme") is False
-    assert api.badge_exists("keepme") is True
-
-    keep_badge = api.get_badge("keepme")
-    assert {tag.name for tag in keep_badge.tags} == {"common-tag"}
-
-
-@pytest.fixture
-def create_badge_with_legacy(api, dummy_issuer_id):
-    def _create_badge(name, legacy=False, desc=None, tags=None):
-        badge_id = api.add_badge(
-            name=name,
-            image=f"{name.lower().replace(' ', '_')}.png",
-            desc=desc or f"{name} description",
-            criteria=f"{name} criteria",
-            issuer_id=dummy_issuer_id,
-            tags=tags,
-        )
-        badge = api.get_badge(badge_id)
-        badge.legacy = legacy
-        api.session.flush()
-        return badge_id
-
-    return _create_badge
-
-
-@pytest.fixture
-def team_with_legacy_badges(api, create_badge_with_legacy):
-    """Create a team with both legacy and non-legacy badges"""
-    team_id = api.create_team("TestTeam")
-    series_id = api.create_series("TestSeries", "A test series", team_id, "test, series")
-
-    legacy_badge = create_badge_with_legacy("Legacy Team Badge", legacy=True)
-    active_badge = create_badge_with_legacy("Active Team Badge")
-
-    api.create_milestone(1, legacy_badge, series_id)
-    api.create_milestone(2, active_badge, series_id)
-
-    return team_id, legacy_badge, active_badge
-
-
-@pytest.fixture
-def tagged_badges(create_badge_with_legacy):
-    """Create badges with tags, both legacy and non-legacy"""
-    legacy_tagged = create_badge_with_legacy(
-        name="Legacy Python Badge",
-        legacy=True,
-        tags="python, legacy, programming",
-    )
-    active_tagged = create_badge_with_legacy(
-        name="Active Python Badge",
-        tags="python, active, programming",
-    )
-
-    return legacy_tagged, active_tagged
-
-
-@pytest.mark.parametrize("include_legacy", [False, True])
-def test_get_all_badges_legacy_filtering(api, create_badge_with_legacy, include_legacy):
-    """Test that get_all_badges respects include_legacy parameter"""
-    legacy_badge = create_badge_with_legacy("Legacy Badge", legacy=True)
-    active_badge = create_badge_with_legacy("Active Badge")
-
-    results = list(api.get_all_badges(include_legacy=include_legacy))
-    badge_ids = [b.id for b in results]
-
-    assert (legacy_badge in badge_ids) is include_legacy
-    assert active_badge in badge_ids
-
-
-@pytest.mark.parametrize("include_legacy", [False, True])
-def test_get_badges_from_tags_legacy_filtering(api, tagged_badges, include_legacy):
-    """Test that get_badges_from_tags respects include_legacy parameter"""
-    legacy_id, active_id = tagged_badges
-
-    results = api.get_badges_from_tags(["python"], include_legacy=include_legacy)
-    badge_ids = [b.id for b in results]
-
-    assert (legacy_id in badge_ids) is include_legacy
-    assert active_id in badge_ids
-
-
-@pytest.mark.parametrize("include_legacy", [False, True])
-def test_get_badges_from_tags_match_all_legacy_filtering(
-    api, create_badge_with_legacy, include_legacy
-):
-    """Test that get_badges_from_tags with match_all=True respects include_legacy"""
-    legacy_multi = create_badge_with_legacy(
-        name="Legacy Multi Tag",
-        legacy=True,
-        tags="python, web, legacy",
-    )
-    active_multi = create_badge_with_legacy(
-        name="Active Multi Tag",
-        tags="python, web, active",
-    )
-
-    results = api.get_badges_from_tags(
-        ["python", "web"], match_all=True, include_legacy=include_legacy
-    )
-    badge_ids = [b.id for b in results]
-
-    assert (legacy_multi in badge_ids) is include_legacy
-    assert active_multi in badge_ids
-
-
-@pytest.mark.parametrize("include_legacy", [False, True])
-def test_get_badges_from_team_legacy_filtering(api, team_with_legacy_badges, include_legacy):
-    """Test that get_badges_from_team respects include_legacy parameter"""
-    team_id, legacy_badge, active_badge = team_with_legacy_badges
-
-    results = api.get_badges_from_team(team_id, include_legacy=include_legacy)
-    badge_ids = [b.id for b in results]
-
-    assert (legacy_badge in badge_ids) is include_legacy
-    assert active_badge in badge_ids
-
-
-@pytest.mark.parametrize("include_legacy", [False, True])
-def test_get_badges_by_string_legacy_filtering(api, create_badge_with_legacy, include_legacy):
-    """Test that get_badges_by_string respects include_legacy parameter"""
-    legacy_search = create_badge_with_legacy(
-        "Legacy SearchBadge",
-        legacy=True,
-    )
-    active_search = create_badge_with_legacy("Active SearchBadge")
-
-    results = api.get_badges_by_string("SearchBadge", include_legacy=include_legacy)
-    badge_ids = [b.id for b in results["badges"]]
-
-    assert (legacy_search in badge_ids) is include_legacy
-    assert active_search in badge_ids
-
-
-@pytest.mark.parametrize("include_legacy", [False, True])
-def test_get_badges_by_string_description_search_legacy_filtering(
-    api, create_badge_with_legacy, include_legacy
-):
-    """Test that get_badges_by_string searches descriptions and respects include_legacy"""
-    legacy_desc = create_badge_with_legacy(
-        name="Badge One",
-        legacy=True,
-        desc="This is a legacy special description",
-    )
-    active_desc = create_badge_with_legacy(
-        name="Badge Two",
-        desc="This is an active special description",
-    )
-
-    results = api.get_badges_by_string("special description", include_legacy=include_legacy)
-    badge_ids = [b.id for b in results["badges"]]
-
-    assert (legacy_desc in badge_ids) is include_legacy
-    assert active_desc in badge_ids
